refactor: Extract wiki domain
