+++ scripts/wiki-content.ts
@@ -7,7 +7,7 @@ import { ArgumentParser } from "argparse";
 import { config } from "dotenv";
 import jscodeshift from "jscodeshift";
 
-import { wikiPageSlugs } from "src/domain/data";
+import { wikiPageSlugs } from "src/domain/wiki";
 import { getCachedWikiPages, WikiPage } from "src/domain/gitlab-wiki-api";
 
 // Load environment variables from .env file
